Wire flow-sensitive referent filter into edge resolution
Shadowed declarations no longer keep their upstream alive at a later
access. The visitor now calls ``live_referents`` for each access with
multiple candidate referents and drops the ones killed by a later
binding on every control-flow path. The seven false-positive cases
from ``test_redeclaration_limitation`` now produce the intended edge
sets and have been promoted into ``test_shadowed_declarations``.
