+++ tests/test_declarations.py
@@ -750,10 +750,9 @@ def test_declarations(build_decl_graph, assert_edges, src, expected_edges):
             def f(): pass
             f()
             """,
-            # Two function nodes; ``f()`` resolves to both referents so
-            # both get kept alive. Only the last has a parent edge.
+            # The flow filter drops the shadowed line-1 def at the call
+            # site, so only the surviving ``mod.f@2:0`` appears.
             {
-                "mod -> mod.f@1:0",
                 "mod -> mod.f@2:0",
                 "mod.f@2:0 -> mod",
             },
@@ -766,7 +765,6 @@ def test_declarations(build_decl_graph, assert_edges, src, expected_edges):
             C()
             """,
             {
-                "mod -> mod.C@1:0",
                 "mod -> mod.C@2:0",
                 "mod.C@2:0 -> mod",
             },
@@ -779,10 +777,10 @@ def test_declarations(build_decl_graph, assert_edges, src, expected_edges):
             def f(): a()
             f()
             """,
-            # The second ``def f`` references ``a`` from its body. That
-            # edge lives on the ``mod.f@3:0`` node only.
+            # ``f()`` only reaches the last ``def f`` after filtering.
+            # The shadowed ``mod.f@2:0`` had no outgoing references
+            # anyway, so it drops out of the graph entirely.
             {
-                "mod -> mod.f@2:0",
                 "mod -> mod.f@3:0",
                 "mod.a@1:0 -> mod",
                 "mod.f@3:0 -> mod",
@@ -800,7 +798,6 @@ def test_declarations(build_decl_graph, assert_edges, src, expected_edges):
             """,
             # Decorator edge lives on the decorated (second) decl only.
             {
-                "mod -> mod.f@2:0",
                 "mod -> mod.f@4:0",
                 "mod.dec@1:0 -> mod",
                 "mod.f@4:0 -> mod",
@@ -872,6 +869,176 @@ def test_declarations(build_decl_graph, assert_edges, src, expected_edges):
 )
 def test_redeclarations(build_decl_graph, assert_positional_edges, src, expected_edges):
     graph = build_decl_graph({"mod.py": src})
+    assert_positional_edges(graph, expected_edges)
+
+
+@pytest.mark.parametrize(
+    "files, expected_edges",
+    [
+        # ------------------------------------------------------------------
+        # Cross-kind shadowing: the flow filter drops the shadowed
+        # referent at the access, so the shadowed decl loses its only
+        # incoming edge and no longer drags its upstream along.
+        # ------------------------------------------------------------------
+        pytest.param(
+            {
+                "other.py": "def f(): pass\n",
+                "mod.py": """
+                from other import f
+                def f(): pass
+                f()
+                """,
+            },
+            # ``f()`` resolves only to the line-2 def; the shadowed
+            # import node at col 18 is still in the graph (its own
+            # import edges wire it to ``other``), but nothing points at
+            # it so ``other`` is no longer kept alive by this module.
+            {
+                "mod -> mod.f@2:0",
+                "mod.f@1:18 -> other",
+                "mod.f@1:18 -> other.f@1:0",
+                "mod.f@2:0 -> mod",
+                "other.f@1:0 -> other",
+            },
+            id="import-shadowed-by-function",
+        ),
+        pytest.param(
+            {
+                "other.py": "def f(): pass\n",
+                "mod.py": """
+                def f(): pass
+                from other import f
+                f()
+                """,
+            },
+            # The shadowing import wins at the call; the shadowed def
+            # at line 1 has no incoming edges and drops from the graph.
+            {
+                "mod -> mod.f@2:18",
+                "mod -> other",
+                "mod -> other.f@1:0",
+                "mod.f@2:18 -> mod",
+                "mod.f@2:18 -> other",
+                "mod.f@2:18 -> other.f@1:0",
+                "other.f@1:0 -> other",
+            },
+            id="function-shadowed-by-import",
+        ),
+        pytest.param(
+            {
+                "other.py": "def f(): pass\n",
+                "mod.py": """
+                from other import f
+                f = 1
+                print(f)
+                """,
+            },
+            {
+                "mod -> mod.f@2:0",
+                "mod.f@1:18 -> other",
+                "mod.f@1:18 -> other.f@1:0",
+                "mod.f@2:0 -> mod",
+                "other.f@1:0 -> other",
+            },
+            id="import-rebound-to-constant",
+        ),
+        pytest.param(
+            {
+                "a.py": "def x(): pass\n",
+                "b.py": "def x(): pass\n",
+                "mod.py": """
+                from a import x
+                from b import x
+                x()
+                """,
+            },
+            # Only the second import is live at the call, so ``a`` is
+            # unreachable from ``mod`` (the line-1 import node lingers
+            # as an orphan but contributes no new incoming edges).
+            {
+                "a.x@1:0 -> a",
+                "b.x@1:0 -> b",
+                "mod -> b",
+                "mod -> b.x@1:0",
+                "mod -> mod.x@2:14",
+                "mod.x@1:14 -> a",
+                "mod.x@1:14 -> a.x@1:0",
+                "mod.x@2:14 -> b",
+                "mod.x@2:14 -> b.x@1:0",
+                "mod.x@2:14 -> mod",
+            },
+            id="two-imports-same-alias-last-wins",
+        ),
+        # ------------------------------------------------------------------
+        # Same-kind redeclaration: the dead body is no longer kept alive
+        # through the shadowed decl.
+        # ------------------------------------------------------------------
+        pytest.param(
+            {
+                "mod.py": """
+                def dead_helper(): pass
+                def f(): dead_helper()
+                def f(): pass
+                f()
+                """,
+            },
+            # ``f()`` only reaches the line-3 def now, so the dead body
+            # at line 2 becomes an orphan -- ``dead_helper`` is no
+            # longer reachable.
+            {
+                "mod -> mod.f@3:0",
+                "mod.dead_helper@1:0 -> mod",
+                "mod.f@2:0 -> mod.dead_helper@1:0",
+                "mod.f@3:0 -> mod",
+            },
+            id="dead-function-body-is-orphaned",
+        ),
+        pytest.param(
+            {
+                "mod.py": """
+                def dead_helper(): pass
+                class C:
+                    def m(self): dead_helper()
+                class C: pass
+                C()
+                """,
+            },
+            {
+                "mod -> mod.C@4:0",
+                "mod.C@2:0 -> mod.dead_helper@1:0",
+                "mod.C@4:0 -> mod",
+                "mod.dead_helper@1:0 -> mod",
+            },
+            id="dead-method-body-is-orphaned",
+        ),
+        pytest.param(
+            {
+                "mod.py": """
+                def a(): pass
+                def b(): pass
+                def f(): a()
+                def f(): b()
+                def f(): pass
+                f()
+                """,
+            },
+            # Only the last ``def f`` is live at the call; the two
+            # earlier bodies keep their body edges but nothing reaches
+            # them.
+            {
+                "mod -> mod.f@5:0",
+                "mod.a@1:0 -> mod",
+                "mod.b@2:0 -> mod",
+                "mod.f@3:0 -> mod.a@1:0",
+                "mod.f@4:0 -> mod.b@2:0",
+                "mod.f@5:0 -> mod",
+            },
+            id="chain-of-shadowed-functions-orphans-earlier-bodies",
+        ),
+    ],
+)
+def test_shadowed_declarations(build_decl_graph, assert_positional_edges, files, expected_edges):
+    graph = build_decl_graph(files)
     assert_positional_edges(graph, expected_edges)
