+++ tests/test_limitations.py
@@ -97,189 +97,3 @@ import pytest
 def test_limitation(build_decl_graph, assert_edges, files, expected_edges):
     graph = build_decl_graph(files)
     assert_edges(graph, expected_edges)
-
-
-@pytest.mark.parametrize(
-    "files, expected_edges",
-    [
-        # ------------------------------------------------------------------
-        # Cross-kind shadowing: both the shadowed and shadowing decl
-        # survive as distinct nodes at the same fqname, so the shadowed
-        # node's upstream edges keep its (dead) import source alive.
-        # ------------------------------------------------------------------
-        pytest.param(
-            {
-                "other.py": "def f(): pass\n",
-                "mod.py": """
-                from other import f
-                def f(): pass
-                f()
-                """,
-            },
-            # The import at col 18 and the ``def`` at col 0 are distinct
-            # nodes. The shadowed import still has outgoing edges to
-            # ``other`` / ``other.f``, and the module-level call reaches
-            # both, so ``other`` stays alive.
-            {
-                "mod -> mod.f@1:18",
-                "mod -> mod.f@2:0",
-                "mod -> other",
-                "mod -> other.f@1:0",
-                "mod.f@1:18 -> other",
-                "mod.f@1:18 -> other.f@1:0",
-                "mod.f@2:0 -> mod",
-                "other.f@1:0 -> other",
-            },
-            id="import-shadowed-by-function-keeps-import-alive",
-        ),
-        pytest.param(
-            {
-                "other.py": "def f(): pass\n",
-                "mod.py": """
-                def f(): pass
-                from other import f
-                f()
-                """,
-            },
-            {
-                "mod -> mod.f@1:0",
-                "mod -> mod.f@2:18",
-                "mod -> other",
-                "mod -> other.f@1:0",
-                "mod.f@2:18 -> mod",
-                "mod.f@2:18 -> other",
-                "mod.f@2:18 -> other.f@1:0",
-                "other.f@1:0 -> other",
-            },
-            id="function-shadowed-by-import-keeps-function-alive",
-        ),
-        pytest.param(
-            {
-                "other.py": "def f(): pass\n",
-                "mod.py": """
-                from other import f
-                f = 1
-                print(f)
-                """,
-            },
-            {
-                "mod -> mod.f@1:18",
-                "mod -> mod.f@2:0",
-                "mod -> other",
-                "mod -> other.f@1:0",
-                "mod.f@1:18 -> other",
-                "mod.f@1:18 -> other.f@1:0",
-                "mod.f@2:0 -> mod",
-                "other.f@1:0 -> other",
-            },
-            id="import-rebound-to-constant-keeps-import-alive",
-        ),
-        pytest.param(
-            {
-                "a.py": "def x(): pass\n",
-                "b.py": "def x(): pass\n",
-                "mod.py": """
-                from a import x
-                from b import x
-                x()
-                """,
-            },
-            # Two distinct ``mod.x`` import nodes, one per source module.
-            # Both stay alive, dragging ``a`` and ``b`` along.
-            {
-                "a.x@1:0 -> a",
-                "b.x@1:0 -> b",
-                "mod -> a",
-                "mod -> a.x@1:0",
-                "mod -> b",
-                "mod -> b.x@1:0",
-                "mod -> mod.x@1:14",
-                "mod -> mod.x@2:14",
-                "mod.x@1:14 -> a",
-                "mod.x@1:14 -> a.x@1:0",
-                "mod.x@2:14 -> b",
-                "mod.x@2:14 -> b.x@1:0",
-                "mod.x@2:14 -> mod",
-            },
-            id="two-imports-same-alias-both-kept-alive",
-        ),
-        # ------------------------------------------------------------------
-        # Same-kind redeclaration: node identity is now per-position, so
-        # each ``def f`` is its own node. The shadowed node still has
-        # outgoing body edges which keep their targets alive -- phase 2
-        # (flow-sensitive referent filtering) should drop the edge
-        # ``mod -> <shadowed>`` at the call site so the dead body
-        # becomes unreachable and collects dead_helper along with it.
-        # ------------------------------------------------------------------
-        pytest.param(
-            {
-                "mod.py": """
-                def dead_helper(): pass
-                def f(): dead_helper()
-                def f(): pass
-                f()
-                """,
-            },
-            # ``mod.f@2:0`` is the shadowed decl; it has the body edge
-            # to ``dead_helper`` and no parent edge (trie only keeps the
-            # last). ``f()`` reaches both f nodes, so the dead body is
-            # kept alive -- false positive.
-            {
-                "mod -> mod.f@2:0",
-                "mod -> mod.f@3:0",
-                "mod.dead_helper@1:0 -> mod",
-                "mod.f@2:0 -> mod.dead_helper@1:0",
-                "mod.f@3:0 -> mod",
-            },
-            id="dead-function-body-kept-alive",
-        ),
-        pytest.param(
-            {
-                "mod.py": """
-                def dead_helper(): pass
-                class C:
-                    def m(self): dead_helper()
-                class C: pass
-                C()
-                """,
-            },
-            {
-                "mod -> mod.C@2:0",
-                "mod -> mod.C@4:0",
-                "mod.C@2:0 -> mod.dead_helper@1:0",
-                "mod.C@4:0 -> mod",
-                "mod.dead_helper@1:0 -> mod",
-            },
-            id="dead-method-body-kept-alive",
-        ),
-        pytest.param(
-            {
-                "mod.py": """
-                def a(): pass
-                def b(): pass
-                def f(): a()
-                def f(): b()
-                def f(): pass
-                f()
-                """,
-            },
-            # Three ``def f`` nodes. The call site reaches all three;
-            # only the last has a parent edge. The body edges to ``a``
-            # and ``b`` live on the shadowed nodes at lines 3 and 4.
-            {
-                "mod -> mod.f@3:0",
-                "mod -> mod.f@4:0",
-                "mod -> mod.f@5:0",
-                "mod.a@1:0 -> mod",
-                "mod.b@2:0 -> mod",
-                "mod.f@3:0 -> mod.a@1:0",
-                "mod.f@4:0 -> mod.b@2:0",
-                "mod.f@5:0 -> mod",
-            },
-            id="chain-of-shadowed-functions-keeps-all-bodies-alive",
-        ),
-    ],
-)
-def test_redeclaration_limitation(build_decl_graph, assert_positional_edges, files, expected_edges):
-    graph = build_decl_graph(files)
-    assert_positional_edges(graph, expected_edges)
